Only update earliest timestamp if its smaller than the existing
